requirements.txt
matplotlib==3.10.5
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1


